requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cssselect==1.1.0
cycler==0.10.0
debugpy==1.5.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.0
flatten-json==0.1.13
huggingface-hub==0.4.0
idna==2.10
ipykernel==6.4.1
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
newspaper3k==0.2.8
nltk==3.6.5
notebook==6.4.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py4j==0.10.7
pyarrow==5.0.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyspark==2.4.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
pyzmq==22.0.3
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.11.10
requests==2.25.1
requests-file==1.5.1
sacremoses==0.0.47
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
terminado==0.10.0
testpath==0.5.0
threadpoolctl==3.0.0
tinysegmenter==0.3
tldextract==3.1.2
tokenizers==0.10.3
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
transformers==4.13.0
typing_extensions==4.0.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
